+++ vkconfig_gui/widget_setting_filesystem.cpp
@@ -228,6 +228,9 @@ void WidgetSettingFilesystem::textFieldReturnPressed() {
         this->LoadPath(path);
     }
 
+    Configurator& configurator = Configurator::Get();
+    ::CheckMessage(configurator.ignored_messages, this->meta.messages, this->data_set);
+
     emit refreshEnableOnly();
 }