+++ vkconfig_gui/widget_setting_int.cpp
@@ -150,6 +150,9 @@ void WidgetSettingInt::OnErrorValue() {
 void WidgetSettingInt::OnValidValue() {
     this->timer_valid->stop();
 
+    Configurator& configurator = Configurator::Get();
+    ::CheckMessage(configurator.ignored_messages, this->meta.messages, this->data_set);
+
     emit refreshEnableOnly();
 }